backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
celery==5.3.6
python-alipay-sdk==3.3.0
requests==2.31.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
freezegun==1.4.0
